+++ installer/test.php
@@ -324,12 +324,8 @@ if (isset($_POST['sendmail'])) {
         // send mail using configured SMTP server
         $CONFIG = $RCI->config;
 
-        if (!empty($_POST['_smtp_user'])) {
-            $CONFIG['smtp_user'] = $_POST['_smtp_user'];
-        }
-        if (!empty($_POST['_smtp_pass'])) {
-            $CONFIG['smtp_pass'] = $_POST['_smtp_pass'];
-        }
+        $CONFIG['smtp_user'] = trim($_POST['_smtp_user']);
+        $CONFIG['smtp_pass'] = trim($_POST['_smtp_pass']);
 
         $mail_object = new Mail_mime();
         $send_headers = $mail_object->headers($headers);
@@ -418,7 +414,7 @@ if (isset($_POST['imaptest']) && !empty($_POST['_host']) && !empty($_POST['_user
     $imap_ssl = false;
 
     $a_host = parse_url($imap_host);
-    if ($a_host['host']) {
+    if (!empty($a_host['host'])) {
         $imap_host = $a_host['host'];
         $imap_ssl = (isset($a_host['scheme']) && in_array($a_host['scheme'], ['ssl', 'imaps', 'tls'])) ? $a_host['scheme'] : null;
         $imap_port = $a_host['port'] ?? ($imap_ssl && $imap_ssl != 'tls' ? 993 : 143);